feat(golang): add delete method for book-library
