+++ .github/workflows/golang-book-library-test.yml
@@ -17,7 +17,7 @@ jobs:
       - name: Setup Go
         uses: actions/setup-go@v3.4.0
         with:
-          go-version: "1.18"
+          go-version: "1.19"
 
       - name: Install dependencies
         working-directory: "golang/book-library"
