+++ golang/book-library/README.md
@@ -1,12 +1,14 @@
 # Book Library
 
 ## Overview
+
 A basic virtual library book management system providing a REST API
 
 ## To Do
+
 - ~~Add tests~~
 - ~~Add Dockerfile~~
-- Add Github action for continuous integration testing
+- ~~Add Github action for continuous integration testing~~
 - Add database (mongodb)
 - Add Kubernetes deployment
 - Add tilt environment for local deployment
